chore: move code around run ci
