+++ npm/react/cypress.config.js
@@ -1,14 +1,17 @@
-module.exports = {
-  'viewportWidth': 400,
-  'viewportHeight': 400,
-  'video': false,
-  'projectId': 'z9dxah',
-  'env': {
-    'reactDevtools': true,
+const { defineConfig } = require('cypress')
+
+module.exports = defineConfig({
+  viewportWidth: 400,
+  viewportHeight: 400,
+  video: false,
+  projectId: 'z9dxah',
+  env: {
+    reactDevtools: true,
   },
-  'experimentalFetchPolyfill': true,
-  'component': {
-    'excludeSpecPattern': [
+  experimentalFetchPolyfill: true,
+  component: {
+    experimentalSingleTabRunMode: true,
+    excludeSpecPattern: [
       '**/__snapshots__/*',
       '**/__image_snapshots__/*',
       'examples/**/*',
@@ -18,4 +21,4 @@ module.exports = {
       bundler: 'vite',
     },
   },
-}
+})
