+++ tooling/v8-snapshot/cache/dev-linux/snapshot-meta.cache.json
@@ -33,6 +33,11 @@
     "./node_modules/mocha/node_modules/debug/src/node.js",
     "./node_modules/morgan/node_modules/debug/src/node.js",
     "./node_modules/prettier/index.js",
+    "./node_modules/prettier/parser-babel.js",
+    "./node_modules/prettier/parser-espree.js",
+    "./node_modules/prettier/parser-flow.js",
+    "./node_modules/prettier/parser-meriyah.js",
+    "./node_modules/prettier/parser-typescript.js",
     "./node_modules/prettier/third-party.js",
     "./node_modules/send/node_modules/debug/src/node.js",
     "./node_modules/stream-parser/node_modules/debug/src/node.js",
@@ -45,14 +50,20 @@
     "./packages/https-proxy/lib/ca.js",
     "./packages/net-stubbing/node_modules/debug/src/node.js",
     "./packages/network/node_modules/minimatch/minimatch.js",
+    "./packages/server/lib/browsers/utils.ts",
+    "./packages/server/lib/cloud/exception.ts",
+    "./packages/server/lib/errors.ts",
     "./packages/server/lib/modes/record.js",
     "./packages/server/lib/modes/run.ts",
     "./packages/server/lib/open_project.ts",
     "./packages/server/lib/project-base.ts",
     "./packages/server/lib/socket-ct.ts",
+    "./packages/server/lib/util/process_profiler.ts",
     "./packages/server/node_modules/@benmalka/foxdriver/node_modules/graceful-fs/polyfills.js",
+    "./packages/server/node_modules/ci-info/index.js",
     "./packages/server/node_modules/glob/node_modules/minimatch/minimatch.js",
     "./packages/server/node_modules/graceful-fs/polyfills.js",
+    "./packages/server/node_modules/is-ci/index.js",
     "./packages/server/node_modules/mocha/node_modules/debug/src/node.js",
     "./packages/server/node_modules/signal-exit/index.js",
     "./process-nextick-args/index.js",
@@ -3425,6 +3436,7 @@
     "./packages/server/node_modules/@benmalka/foxdriver/node_modules/graceful-fs/legacy-streams.js",
     "./packages/server/node_modules/@benmalka/foxdriver/package.json",
     "./packages/server/node_modules/ansi-regex/index.js",
+    "./packages/server/node_modules/ci-info/vendors.json",
     "./packages/server/node_modules/cli-table3/index.js",
     "./packages/server/node_modules/cli-table3/src/cell.js",
     "./packages/server/node_modules/cli-table3/src/layout-manager.js",
@@ -3529,5 +3541,5 @@
     "./tooling/v8-snapshot/cache/dev-linux/snapshot-entry.js"
   ],
   "deferredHashFile": "yarn.lock",
-  "deferredHash": "6e64b278585907ecd02c6688a3964ef282453f2ca9753603ade51dc14ab6f7ee"
+  "deferredHash": "c081f05e37cf807c0129174be65a0187071f7a9f67354fae7aeaf8767d9608e3"
 }
